+++ market-generator/src/main/java/com/market/banica/generator/service/task/SnapshotPersistenceTask.java
